+++ com.unity.shadergraph/Editor/Generation/Processors/PropertyCollector.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using UnityEditor.ShaderGraph.Internal;
+using TextureDimension = UnityEngine.Rendering.TextureDimension;
 
 namespace UnityEditor.ShaderGraph
 {
@@ -11,6 +11,7 @@ namespace UnityEditor.ShaderGraph
         {
             public string name;
             public int textureId;
+            public TextureDimension dimension;
             public bool modifiable;
         }
 
@@ -204,11 +205,11 @@ namespace UnityEditor.ShaderGraph
 #endif
         }
 
-        public List<TextureInfo> GetConfiguredTexutres()
+        public List<TextureInfo> GetConfiguredTextures()
         {
             var result = new List<TextureInfo>();
 
-            // TODO: this should be interface based instead of looking for hard codeded tyhpes
+            // TODO: this should be interface based instead of looking for hard coded types
 
             foreach (var prop in properties.OfType<Texture2DShaderProperty>())
             {
@@ -218,6 +219,7 @@ namespace UnityEditor.ShaderGraph
                     {
                         name = prop.referenceName,
                         textureId = prop.value.texture != null ? prop.value.texture.GetInstanceID() : 0,
+                        dimension = TextureDimension.Tex2D,
                         modifiable = prop.modifiable
                     };
                     result.Add(textureInfo);
@@ -232,6 +234,7 @@ namespace UnityEditor.ShaderGraph
                     {
                         name = prop.referenceName,
                         textureId = prop.value.textureArray != null ? prop.value.textureArray.GetInstanceID() : 0,
+                        dimension = TextureDimension.Tex2DArray,
                         modifiable = prop.modifiable
                     };
                     result.Add(textureInfo);
@@ -246,6 +249,7 @@ namespace UnityEditor.ShaderGraph
                     {
                         name = prop.referenceName,
                         textureId = prop.value.texture != null ? prop.value.texture.GetInstanceID() : 0,
+                        dimension = TextureDimension.Tex3D,
                         modifiable = prop.modifiable
                     };
                     result.Add(textureInfo);
@@ -260,6 +264,7 @@ namespace UnityEditor.ShaderGraph
                     {
                         name = prop.referenceName,
                         textureId = prop.value.cubemap != null ? prop.value.cubemap.GetInstanceID() : 0,
+                        dimension = TextureDimension.Cube,
                         modifiable = prop.modifiable
                     };
                     result.Add(textureInfo);